moved config package
